twitter-stock-prediction/requirements.txt
pandas
numpy
kafka-python
vaderSentiment
scikit-learn
matplotlib
streamlit
pyspark
yfinance